backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
pydantic-settings==2.4.0
google-cloud-compute==1.19.0
google-cloud-dns==0.36.0
google-cloud-resource-manager==1.12.0
google-cloud-storage==2.18.0
google-auth==2.34.0
asyncssh==2.17.0
python-multipart==0.0.9
httpx==0.27.0
aiofiles==24.1.0
